update e2e tests for EBS options
We noticed a ValidationException occurred in our test case because
EBSOptions wasn't specified and is required for the default AES
instance type. In addition to telling the code generator about this
TerminalCondition, we had to specify the 404 for ElastisearchDomain was
ResourceNotFoundException because the coral model has the code as 409.

Finally, after running a few iterations of the test, we set timeouts and
waits appropriately for checking the creation/deletion events for an AES
domain.