clean up e2e smoke test for ES Domain
Updates the e2e test for creating, reading and deleting an Elasticsearch
Domain in a few ways to help stabilize the test. We now check for the
DomainStatus.Processing field to go from True -> False after creating
the ES Domain.

Similarly, when we delete the CR representing the ES Domain from
Kubernetes, we now poll for the ES Domain and error out if the
DomainStatus.Deleted field is False when returned from
DescribeElasticsearchDomain.

The simple e2e test smoke test takes around 830s (roughly 12 minutes)
for me, pretty consistently. When speaking with Sushant from the AES
team, he mentioned those times are inline with expectations and that
there were no changes to instance type or EBS options that would reduce
the amount of time for creation or deletion.